Refactor code into multiple packages under the duke package
The classes are all unorganised inside one directory.

Divide into multiple packages as the number of classes.

This ensures better organisation of code.
