+++ data/tasks.txt
@@ -1,3 +1,4 @@
-T;haha; 
+T;haha;X
 D;yo;X;2020-02-02
 T;andre; 
+T;commit to github;
