+++ src/main/java/drake/DrakeException.java
@@ -1,3 +1,5 @@
+package drake;
+
 public class DrakeException extends Exception {
     public DrakeException(String errorMessage) {
         super(errorMessage);
