+++ src/main/java/drake/EmptyDescriptionException.java
@@ -1,3 +1,5 @@
+package drake;
+
 public class EmptyDescriptionException extends DrakeException {
     public EmptyDescriptionException() {
         super("Well well well! A task without a description is like a picture of Singapore without MBS");
