+++ src/main/java/drake/IncompatibleCommandException.java
@@ -1,7 +1,6 @@
+package drake;
+
 public class IncompatibleCommandException extends DrakeException {
-    public IncompatibleCommandException() {
-        super("These two go together like oil mixes with water");
-    }
 
     public IncompatibleCommandException(String error) {
         super("These two go together like oil mixes with water. " + error);
