+++ src/main/java/drake/InvalidTaskNumberException.java
@@ -1,3 +1,5 @@
+package drake;
+
 public class InvalidTaskNumberException extends IncompatibleCommandException {
     public InvalidTaskNumberException() {
         super("That task number doesn't exist!");
