+++ src/main/java/drake/UnknownCommandException.java
@@ -1,3 +1,5 @@
+package drake;
+
 public class UnknownCommandException extends DrakeException {
     public UnknownCommandException() {
         super("Uh oh spaghettios I don't know what that means!");
