+++ src/main/java/drake/commands/ByeCommand.java
@@ -1,3 +1,10 @@
+package drake.commands;
+
+import drake.DrakeException;
+import drake.Storage;
+import drake.TaskList;
+import drake.Ui;
+
 import java.io.IOException;
 
 public class ByeCommand extends Command {
