+++ src/main/java/drake/commands/Command.java
@@ -1,3 +1,10 @@
+package drake.commands;
+
+import drake.DrakeException;
+import drake.Storage;
+import drake.TaskList;
+import drake.Ui;
+
 import java.io.IOException;
 
 public abstract class Command {
