+++ src/main/java/drake/commands/CommandType.java
@@ -1,3 +1,5 @@
+package drake.commands;
+
 public enum CommandType {
     MARK,
     UNMARK,
