+++ src/main/java/drake/commands/EventCommand.java
@@ -1,12 +1,18 @@
+package drake.commands;
+
+import drake.*;
+import drake.tasks.Event;
+import drake.tasks.Task;
+
 import java.io.IOException;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-public class DeadlineCommand extends CreateTaskCommand {
+public class EventCommand extends CreateTaskCommand {
 
-    private final static Pattern descriptionPattern = Pattern.compile("(?<taskName>.*) /by (?<by>.*)");
+    private final static Pattern descriptionPattern = Pattern.compile("(?<taskName>.*) /at (?<at>.*)");
 
-    public DeadlineCommand(String fullInput) {
+    public EventCommand(String fullInput) {
         super(fullInput);
     }
 
@@ -14,10 +20,10 @@ public class DeadlineCommand extends CreateTaskCommand {
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DrakeException, IOException {
         Matcher match = descriptionPattern.matcher(description);
         if (!match.matches()) {
-            throw new IncompatibleCommandException("A deadline task without a deadline?");
+            throw new IncompatibleCommandException("An event task without an event time?");
         }
         ui.printLine("I've added this task:");
-        Task addedTask = tasks.addTask(new Deadline(match.group("taskName"), match.group("by")));
+        Task addedTask = tasks.addTask(new Event(match.group("taskName"), match.group("at")));
         ui.printLine(addedTask);
         storage.addTask(addedTask);
         super.execute(tasks, ui, storage);
