+++ src/main/java/drake/commands/MarkCommand.java
@@ -1,17 +1,21 @@
+package drake.commands;
+
+import drake.*;
+
 import java.io.IOException;
 
-public class UnmarkCommand extends TaskOperationCommand {
+public class MarkCommand extends TaskOperationCommand {
 
-    public UnmarkCommand(String fullInput) throws IncompatibleCommandException {
+    public MarkCommand(String fullInput) throws IncompatibleCommandException {
         super(fullInput);
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DrakeException, IOException {
         if (tasks.isValidTaskNumber(taskNumber)) {
-            System.out.println("I've marked this task as not done (yet ;))");
-            tasks.markAsNotDone(taskNumber);
-            storage.updateTask(taskNumber, CommandType.UNMARK);
+            System.out.println("I've marked this task as done!");
+            tasks.markAsDone(taskNumber);
+            storage.updateTask(taskNumber, CommandType.MARK);
             ui.printLine(tasks.getTaskToString(taskNumber));
         } else {
             throw new InvalidTaskNumberException();
