+++ src/main/java/drake/commands/UnmarkCommand.java
@@ -1,18 +1,22 @@
+package drake.commands;
+
+import drake.*;
+
 import java.io.IOException;
 
-public class DeleteCommand extends TaskOperationCommand {
+public class UnmarkCommand extends TaskOperationCommand {
 
-    public DeleteCommand(String fullInput) throws IncompatibleCommandException {
+    public UnmarkCommand(String fullInput) throws IncompatibleCommandException {
         super(fullInput);
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DrakeException, IOException {
         if (tasks.isValidTaskNumber(taskNumber)) {
-            System.out.println("I've removed this task: ");
+            System.out.println("I've marked this task as not done (yet ;))");
+            tasks.markAsNotDone(taskNumber);
+            storage.updateTask(taskNumber, CommandType.UNMARK);
             ui.printLine(tasks.getTaskToString(taskNumber));
-            tasks.removeTask(taskNumber);
-            storage.updateTask(taskNumber, CommandType.DELETE);
         } else {
             throw new InvalidTaskNumberException();
         }
