+++ src/main/java/drake/tasks/Deadline.java
@@ -1,3 +1,5 @@
+package drake.tasks;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
