+++ text-ui-test/EXPECTED.TXT
@@ -6,7 +6,7 @@ DRAKE DRAKE DRAKE DRAKE DRAKE
 DRAKE DRAKE DRAKE DRAKE DRAKE
 DRAKE DRAKE DRAKE DRAKE DRAKE
 !@#$%^&*()-+!@#$%^&*()`~`!@#$
-Drake's (me) the kind of guy to help you out uwu
+drake.Drake's (me) the kind of guy to help you out uwu
 Go ahead, make that hotline bling
 ------------------------------------------------------
 ------------------------------------------------------
